@@ -5,7 +5,7 @@
     "description": "TODO",
     "scripts": {
         "build": "webpack",
-        "test": "jest --verbose",
+        "test": "NODE_OPTIONS=--experimental-vm-modules jest --verbose",
         "lint": "npx eslint ."
     },
     "repository": {
@@ -22,23 +22,27 @@
     "homepage": "https://github.com/maunzCache/fohacker#README.md",
     "engines": {
         "node": ">=18",
-        "npm": ">=8"
+        "npm": ">=9"
     },
     "os": [
         "linux"
     ],
     "devDependencies": {
-        "eslint": "^8.22.0",
-        "jest": "28.1.3",
-        "jest-environment-jsdom": "28.1.3",
-        "jsdom": "20.0.0",
-        "webpack": "5.74.0",
-        "webpack-cli": "4.10.0"
+        "eslint": "^8.38.0",
+        "jest": "29.5.0",
+        "jest-environment-jsdom": "29.5.0",
+        "jsdom": "21.1.1",
+        "webpack": "5.78.0",
+        "webpack-cli": "5.0.1"
     },
     "dependencies": {
-        "jquery": "3.6.0"
+        "jquery": "3.6.4"
     },
     "jest": {
-        "testEnvironment": "jsdom"
+        "testEnvironment": "jsdom",
+        "transform": {},
+        "roots": [
+            "tests"
+        ]
     }
 }
